requirements.txt
django == 2.1
djangorestframework